+++ app-backend/datamodel/src/main/scala/Scene.scala
@@ -154,6 +154,10 @@ final case class Scene(
     sceneOrder,
     this.metadataFields
   )
+
+  def bucketAndKey: Option[(String, String)] = ingestLocation map {
+    uriToBucketAndKey _
+  }
 }
 
 object Scene {