+++ app-backend/db/src/main/scala/UploadDao.scala
@@ -175,4 +175,8 @@ object UploadDao extends Dao[Upload] {
       }
     }
   }
+
+  def findForAnnotationProject(
+      annotationProjectId: UUID): ConnectionIO[List[Upload]] =
+    query.filter(fr"annotation_project_id = $annotationProjectId").list
 }